Hack: ignore filename for path equality
In order to be able to see/exclude pictures from a specific directory.
This feels more natural, since the file path uses a directory as a filter by default.

Could be otherwise implemented by upping schema version and saving another column for directory path in db, but it's more duplicate data, and would prevent going back.